Now Agda highlighting in Org-Agda works, but Org doesn't :(
Still better this way.

The solution was to turn off font-lock mode; Agda mode does this
already, but I believe Polymode was turning it back on in the Agda
inner mode. And unless it is off in both the host and inner mode,
it seems to turn back on when switching between them.

It's unfortunate that I had to turn it off, as it means we lose
Org fontification. But maybe there's a way that can be applied
in just the Org parts of the buffer?