A reset of my Emacs init, cleaning and reorganising the whole file.
The file is now organised based on the importance of the settings,
to make debugging and refactoring simpler.

This is a work in process. At this time, I have used these settings
in their current state for a few weeks. Some things are still to be
moved over at a later time, but this is functional enough to begin with.