Clean up and simplify some headings and other small details.
In particular, remove uses of fixed-width emphasis and quotes
from headings.

The quotes are not really necessary and add clutter
at the heading level.

Fixed width emphasis sometimes misbehaves on Github's rendering
of my exported Org README;
in particular when the emphasised text is at the start or end
of the heading, the table of contents entry shows the `~` characters.